test(cli): split OCI load unit coverage
